+++ internal/entity/order.go
@@ -6,13 +6,13 @@ import (
 
 // Order struct holds entity of order
 type Order struct {
-	ID         int64     `json:"id"`
-	UserID     int32     `json:"user_id"`
-	BookID     int32     `json:"book_id"`
-	Quantity   int32     `json:"quantity"`
-	Price      int32     `json:"price"`
-	Fee        int32     `json:"fee"`
-	TotalPrice int32     `json:"total_price"`
+	ID         int       `json:"id"`
+	UserID     int       `json:"user_id"`
+	BookID     int       `json:"book_id"`
+	Quantity   int       `json:"quantity"`
+	Price      int       `json:"price"`
+	Fee        int       `json:"fee"`
+	TotalPrice int       `json:"total_price"`
 	CreatedAt  time.Time `json:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at"`
 }
